Moving DiskManager to internal
