+++ android-sdk/src/test/java/co/elastic/apm/android/sdk/internal/features/persistence/DiskManagerTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.features.persistence.disk;
+package co.elastic.apm.android.sdk.internal.features.persistence;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
